+++ tests/db_updates.rs
@@ -5,10 +5,10 @@ use rusqlite::Connection;
 use std::fs;
 
 fn check_test_db(file_name: &str) {
-    if let Ok(metadata) = fs::metadata(file_name) {
-        if metadata.is_file() {
-            fs::remove_file(file_name).expect("Failed to delete existing file");
-        }
+    if let Ok(metadata) = fs::metadata(file_name)
+        && metadata.is_file()
+    {
+        fs::remove_file(file_name).expect("Failed to delete existing file");
     }
 }
